requirements.txt
# pip install -r requirements.txt
numpy
matplotlib
tensorflow>=2.10.*
opencv-python
PyYAML>=5.3.1
tensorflow-hub
pyyaml-include
mailtrap
